базовый роутинг на фронте
