+++ frontend/config/webpack.common.js
@@ -52,47 +52,3 @@ module.exports = (env) => {
     },
   };
 };
-
-// module.exports = {
-//   entry: './src/index.tsx',
-//   module: {
-//     rules: [
-//       {
-//         test: /\.tsx?$/,
-//         use: 'ts-loader',
-//         exclude: /node_modules/,
-//       },
-//       {
-//         test: /\.scss$/,
-//         use: [
-//           MiniCssExtractPlugin.loader,
-//           {
-//             loader: 'css-loader',
-//             options: {
-//               modules: true,
-//             },
-//           },
-//           'sass-loader',
-//         ],
-//       },
-//     ],
-//   },
-//   plugins: [
-//     new HtmlWebpackPlugin({ template: 'public/index.html', title: 'test' }),
-//     new MiniCssExtractPlugin({
-//       filename: '[name].css',
-//       chunkFilename: '[id].css',
-//     }),
-//   ],
-//   resolve: {
-//     alias: {
-//       '@components': path.resolve(__dirname, '..', 'src/components/'),
-//     },
-//     extensions: ['.tsx', '.ts', '.js'],
-//   },
-//   output: {
-//     filename: '[name].bundle.js',
-//     path: path.resolve(__dirname, '..', 'dist'),
-//     clean: true,
-//   },
-// };
