+++ frontend/package.json
@@ -12,7 +12,8 @@
     "css-modules-typescript-loader": "4.0.1",
     "mini-css-extract-plugin": "2.7.6",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@types/node": "20.8.6",
