+++ frontend/src/App/global.module.scss
@@ -1,0 +1,9 @@
+body {
+  width: 100%;
+  height: 100%;
+}
+
+#root {
+  //height: 100%;
+  background-color: red;
+}
