--- frontend/src/components/ListForm/index.ts
@@ -1,1 +1,0 @@
-export * from './ListForm';
