+++ frontend/src/index.tsx
@@ -1,5 +1,7 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { App } from './App/App';
+import * as React from 'react';
+import { createRoot } from 'react-dom/client';
+import { App } from './App';
 
-render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
