+++ frontend/src/routes/LoginPage/LoginPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const LoginPage = () => {
+  return (
+    <h1>Login Page</h1>
+  );
+};
