+++ frontend/src/routes/LoginPage/index.ts
@@ -1,0 +1,1 @@
+export * from "./LoginPage"
