+++ frontend/src/routes/Tests/index.ts
@@ -1,0 +1,1 @@
+export * from './TestsPage';
